organize subsystem folders
