--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzHood/HoodIOSim.java
@@ -1,11 +1,0 @@
-package frc.robot.CatzSubsystems.CatzHood;
-
-import frc.robot.CatzAbstractions.io.GenericIOSim;
-import frc.robot.Utilities.MotorUtil.Gains;
-
-public class HoodIOSim extends GenericIOSim<HoodIO.HoodIOInputs> implements HoodIO{
-
-    public HoodIOSim(Gains gains) {
-        super(gains);
-    }
-}
