--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzHood/HoodIOTalonFX.java
@@ -1,9 +1,0 @@
-package frc.robot.CatzSubsystems.CatzHood;
-
-import frc.robot.CatzAbstractions.io.GenericTalonFXIOReal;
-
-public class HoodIOTalonFX extends GenericTalonFXIOReal<HoodIO.HoodIOInputs> implements HoodIO{
-    public HoodIOTalonFX(MotorIOTalonFXConfig config){
-        super(config);
-    }
-}
