--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzIndexer/IndexerIOSim.java
@@ -1,11 +1,0 @@
-package frc.robot.CatzSubsystems.CatzIndexer;
-
-import frc.robot.CatzAbstractions.io.GenericIOSim;
-import frc.robot.Utilities.MotorUtil.Gains;
-
-public class IndexerIOSim extends GenericIOSim<IndexerIO.IndexerIOInputs> implements IndexerIO{
-
-    public IndexerIOSim(Gains gains) {
-        super(gains);
-    }
-}
