--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzIndexer/IndexerIOTalonFX.java
@@ -1,9 +1,0 @@
-package frc.robot.CatzSubsystems.CatzIndexer;
-
-import frc.robot.CatzAbstractions.io.GenericTalonFXIOReal;
-
-public class IndexerIOTalonFX extends GenericTalonFXIOReal<IndexerIO.IndexerIOInputs> implements IndexerIO{
-    public IndexerIOTalonFX(MotorIOTalonFXConfig config){
-        super(config);
-    }
-}
