+++ SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzIntake/CatzIntakeDeploy/IntakeDeployIO.java
@@ -1,0 +1,13 @@
+package frc.robot.CatzSubsystems.CatzIntake.CatzIntakeDeploy;
+
+import org.littletonrobotics.junction.AutoLog;
+
+import frc.robot.CatzAbstractions.io.GenericMotorIO;
+
+public interface IntakeDeployIO extends GenericMotorIO<IntakeDeployIO.IntakeDeployIOInputs>{
+
+    @AutoLog
+    public static class IntakeDeployIOInputs extends GenericMotorIO.MotorIOInputs{
+
+    }
+}
