+++ SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzIntake/CatzIntakeRoller/IntakeRollerIO.java
@@ -1,0 +1,13 @@
+package frc.robot.CatzSubsystems.CatzIntake.CatzIntakeRoller;
+
+import org.littletonrobotics.junction.AutoLog;
+
+import frc.robot.CatzAbstractions.io.GenericMotorIO;
+
+public interface IntakeRollerIO extends GenericMotorIO<IntakeRollerIO.IntakeRollerIOInputs>{
+
+    @AutoLog
+    public static class IntakeRollerIOInputs extends GenericMotorIO.MotorIOInputs{
+
+    }
+}
