+++ SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzIntake/CatzIntakeRoller/IntakeRollerIOSim.java
@@ -1,0 +1,10 @@
+package frc.robot.CatzSubsystems.CatzIntake.CatzIntakeRoller;
+
+import frc.robot.CatzAbstractions.io.GenericIOSim;
+import frc.robot.Utilities.MotorUtil.Gains;
+
+public class IntakeRollerIOSim extends GenericIOSim<IntakeRollerIO.IntakeRollerIOInputs> implements IntakeRollerIO{
+    public IntakeRollerIOSim(Gains gains){
+        super(gains);
+    }
+}
