--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzIntakeDeploy/IntakeDeployIO.java
@@ -1,13 +1,0 @@
-package frc.robot.CatzSubsystems.CatzIntakeDeploy;
-
-import org.littletonrobotics.junction.AutoLog;
-
-import frc.robot.CatzAbstractions.io.GenericMotorIO;
-
-public interface IntakeDeployIO extends GenericMotorIO<IntakeDeployIO.IntakeDeployIOInputs>{
-
-    @AutoLog
-    public static class IntakeDeployIOInputs extends GenericMotorIO.MotorIOInputs{
-
-    }
-}
