--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzIntakeDeploy/IntakeDeployIOSim.java
@@ -1,10 +1,0 @@
-package frc.robot.CatzSubsystems.CatzIntakeDeploy;
-
-import frc.robot.CatzAbstractions.io.GenericIOSim;
-import frc.robot.Utilities.MotorUtil.Gains;
-
-public class IntakeDeployIOSim extends GenericIOSim<IntakeDeployIO.IntakeDeployIOInputs> implements IntakeDeployIO{
-    public IntakeDeployIOSim(Gains gains){
-        super(gains);
-    }
-}
