--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzIntakeRoller/IntakeRollerIOTalonFX.java
@@ -1,10 +1,0 @@
-package frc.robot.CatzSubsystems.CatzIntakeRoller;
-
-import frc.robot.CatzAbstractions.io.GenericTalonFXIOReal;
-
-public class IntakeRollerIOTalonFX extends GenericTalonFXIOReal<IntakeRollerIO.IntakeRollerIOInputs> implements IntakeRollerIO{
-    public IntakeRollerIOTalonFX(MotorIOTalonFXConfig config){
-        super(config);
-    }
-
-}
