+++ SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzShooter/CatzFlywheels/FlywheelsIO.java
@@ -1,0 +1,12 @@
+package frc.robot.CatzSubsystems.CatzShooter.CatzFlywheels;
+
+import org.littletonrobotics.junction.AutoLog;
+
+import frc.robot.CatzAbstractions.io.GenericMotorIO;
+
+public interface FlywheelsIO extends GenericMotorIO<FlywheelsIO.FlywheelsIOInputs>{
+    @AutoLog
+    public static class FlywheelsIOInputs extends GenericMotorIO.MotorIOInputs{
+
+    }
+}
