+++ SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzShooter/CatzFlywheels/FlywheelsIOTalonFX.java
@@ -1,0 +1,9 @@
+package frc.robot.CatzSubsystems.CatzShooter.CatzFlywheels;
+
+import frc.robot.CatzAbstractions.io.GenericTalonFXIOReal;
+
+public class FlywheelsIOTalonFX extends GenericTalonFXIOReal<FlywheelsIO.FlywheelsIOInputs> implements FlywheelsIO{
+    public FlywheelsIOTalonFX(MotorIOTalonFXConfig config){
+        super(config);
+    }
+}
