+++ SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzShooter/CatzHood/HoodIOSim.java
@@ -1,0 +1,11 @@
+package frc.robot.CatzSubsystems.CatzShooter.CatzHood;
+
+import frc.robot.CatzAbstractions.io.GenericIOSim;
+import frc.robot.Utilities.MotorUtil.Gains;
+
+public class HoodIOSim extends GenericIOSim<HoodIO.HoodIOInputs> implements HoodIO{
+
+    public HoodIOSim(Gains gains) {
+        super(gains);
+    }
+}
