+++ SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzShooter/CatzHood/HoodIOTalonFX.java
@@ -1,0 +1,9 @@
+package frc.robot.CatzSubsystems.CatzShooter.CatzHood;
+
+import frc.robot.CatzAbstractions.io.GenericTalonFXIOReal;
+
+public class HoodIOTalonFX extends GenericTalonFXIOReal<HoodIO.HoodIOInputs> implements HoodIO{
+    public HoodIOTalonFX(MotorIOTalonFXConfig config){
+        super(config);
+    }
+}
