+++ SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzShooter/CatzTurret/TurretIO.java
@@ -1,0 +1,13 @@
+package frc.robot.CatzSubsystems.CatzShooter.CatzTurret;
+
+import org.littletonrobotics.junction.AutoLog;
+
+import frc.robot.CatzAbstractions.io.GenericMotorIO;
+
+public interface TurretIO extends GenericMotorIO<TurretIO.TurretIOInputs>{
+    @AutoLog
+    public static class TurretIOInputs extends GenericMotorIO.MotorIOInputs{}
+
+
+    // public default void runMotor() {}
+}
