+++ SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzShooter/CatzTurret/TurretIOSim.java
@@ -1,0 +1,10 @@
+package frc.robot.CatzSubsystems.CatzShooter.CatzTurret;
+
+import frc.robot.CatzAbstractions.io.GenericIOSim;
+import frc.robot.Utilities.MotorUtil.Gains;
+
+public class TurretIOSim extends GenericIOSim<TurretIO.TurretIOInputs> implements TurretIO{
+    public TurretIOSim(Gains gains){
+        super(gains);
+    }
+}
