--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzShooter/FlywheelsIO.java
@@ -1,12 +1,0 @@
-package frc.robot.CatzSubsystems.CatzShooter;
-
-import org.littletonrobotics.junction.AutoLog;
-
-import frc.robot.CatzAbstractions.io.GenericMotorIO;
-
-public interface FlywheelsIO extends GenericMotorIO<FlywheelsIO.FlywheelsIOInputs>{
-    @AutoLog
-    public static class FlywheelsIOInputs extends GenericMotorIO.MotorIOInputs{
-
-    }
-}
