--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzShooter/FlywheelsIOSim.java
@@ -1,10 +1,0 @@
-package frc.robot.CatzSubsystems.CatzShooter;
-
-import frc.robot.CatzAbstractions.io.GenericIOSim;
-import frc.robot.Utilities.MotorUtil.Gains;
-
-public class FlywheelsIOSim extends GenericIOSim<FlywheelsIO.FlywheelsIOInputs> implements FlywheelsIO{
-    public FlywheelsIOSim(Gains gains) {
-        super(gains);
-    }
-}
