--- SeasonCode2026/src/main/java/frc/robot/CatzSubsystems/CatzShooter/FlywheelsIOTalonFX.java
@@ -1,9 +1,0 @@
-package frc.robot.CatzSubsystems.CatzShooter;
-
-import frc.robot.CatzAbstractions.io.GenericTalonFXIOReal;
-
-public class FlywheelsIOTalonFX extends GenericTalonFXIOReal<FlywheelsIO.FlywheelsIOInputs> implements FlywheelsIO{
-    public FlywheelsIOTalonFX(MotorIOTalonFXConfig config){
-        super(config);
-    }
-}
